# Feature: cache validation rules