--- a/src/LighthouseServiceProvider.php
+++ b/src/LighthouseServiceProvider.php
@@ -101,7 +101,7 @@ class LighthouseServiceProvider extends ServiceProvider
             }
         });
 
-        //$this->app->bind(ProvidesValidationRules::class, ValidationRulesProvider::class);
+        // $this->app->bind(ProvidesValidationRules::class, ValidationRulesProvider::class);
         $this->app->bind(ProvidesValidationRules::class, CacheableValidationRulesProvider::class);
 
         $this->commands(self::COMMANDS);