--- a/src/Schema/AST/DocumentAST.php
+++ b/src/Schema/AST/DocumentAST.php
@@ -46,6 +46,8 @@ class DocumentAST implements Arrayable
 
     public const SCHEMA_EXTENSIONS = 'schemaExtensions';
 
+    public const HASH = 'hash';
+
     /**
      * The types within the schema.
      *
@@ -88,6 +90,11 @@ class DocumentAST implements Arrayable
     /** @var array<int,SchemaExtensionNode> */
     public array $schemaExtensions = [];
 
+    /**
+     * A hash of the schema.
+     */
+    public ?string $hash = null;
+
     /** Create a new DocumentAST instance from a schema. */
     public static function fromSource(string $schema): self
     {
@@ -104,6 +111,7 @@ class DocumentAST implements Arrayable
         }
 
         $instance = new static();
+        $instance->hash = hash('sha256', $schema);
 
         foreach ($documentNode->definitions as $definition) {
             if ($definition instanceof TypeDefinitionNode) {
@@ -195,6 +203,7 @@ class DocumentAST implements Arrayable
             self::DIRECTIVES => array_map([AST::class, 'toArray'], $this->directives),
             self::CLASS_NAME_TO_OBJECT_TYPE_NAME => $this->classNameToObjectTypeNames,
             self::SCHEMA_EXTENSIONS => array_map([AST::class, 'toArray'], $this->schemaExtensions),
+            self::HASH => $this->hash,
         ];
     }
 
@@ -231,6 +240,7 @@ class DocumentAST implements Arrayable
             self::DIRECTIVES => $directives,
             self::CLASS_NAME_TO_OBJECT_TYPE_NAME => $this->classNameToObjectTypeNames,
             self::SCHEMA_EXTENSIONS => $schemaExtensions,
+            self::HASH => $this->hash,
         ] = $ast;
 
         // Utilize the NodeList for lazy unserialization for performance gains.