--- a/src/Schema/AST/DocumentAST.php
+++ b/src/Schema/AST/DocumentAST.php
@@ -32,6 +32,7 @@ use Nuwave\Lighthouse\Support\Utils;
  *     directives: array<int, array<string, mixed>>,
  *     classNameToObjectTypeName: ClassNameToObjectTypeName,
  *     schemaExtensions: array<int, array<string, mixed>>,
+ *     hash: string|null,
  * }
  *
  * @implements \Illuminate\Contracts\Support\Arrayable<string, mixed>