--- a/src/Support/Contracts/ProvidesCacheableValidationRules.php
+++ b/src/Support/Contracts/ProvidesCacheableValidationRules.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+
+namespace Nuwave\Lighthouse\Support\Contracts;
+
+interface ProvidesCacheableValidationRules extends ProvidesValidationRules
+{
+    /**
+     * A set of rules for the first query validation step.
+     *
+     * These rules would be executed first and their result would be cached.
+     *
+     * @return array<string, \GraphQL\Validator\Rules\ValidationRule>
+     */
+    public function cacheableValidationRules(): array;
+
+    /**
+     * A set of rules for the second query validation step.
+     *
+     * These rules would always be executed and not cached.
+     *
+     * Returning `null` enables all available rules.
+     * Empty array skips query validation entirely.
+     *
+     * @return array<string, \GraphQL\Validator\Rules\ValidationRule>|null
+     */
+    public function validationRules(): ?array;
+}