… CHANGELOG.md
… benchmarks/HugeRequestBench.php
… benchmarks/HugeResponseBench.php
… benchmarks/QueryBench.php
… docs/6/performance/query-caching.md
… docs/6/security/validation.md
… docs/master/performance/query-caching.md
… docs/master/security/validation.md
… src/Support/Contracts/ProvidesValidationRules.php
… src/lighthouse.php
… tests/Integration/QueryCachingTest.php
… tests/Integration/ValidationCachingTest.php
… tests/Unit/Schema/AST/DocumentASTTest.php
… tests/Unit/Schema/SchemaBuilderTest.php