resolves [#2018](https://github.com/nuwave/lighthouse/issues/2018)
 
- [x] Added or updated tests
- [x] Documented user facing changes
- [x] Updated CHANGELOG.md

**Changes**

Added possibility to cache query validation successful result. So the 2nd time the query would appear, most validator won't run against it.

Notes:
- added `hash` to `\Nuwave\Lighthouse\Schema\AST\DocumentAST`
- created interface `ProvidesCacheableValidationRules` which extends `ProvidesValidationRules` for backward compatibility. In the next release we might remove `ProvidesValidationRules`. 
- some execution paths doesn't provide query hash, so they won't use this feature
  - `Defer::executeDeferred`
  - `SubscriptionBroadcaster::broadcast`
  - `AsyncMutation::handle`
- validation cache is disabled by default for now. 
- it only caches successful results as `Errors` might fail to serialize.
- passing hashes with schema and query is quite ugly https://github.com/nuwave/lighthouse/blob/0d8ea1665f5ded77ad7ba6dde2aec8c703e6b368/src/GraphQL.php#L393-L397 but these objects are created by `webonyx/graphql-php`, so there is no easy way to extend them and store hashes inside objects.

I also added different benchmarks. `HugeRequestBench` creates requests with a lot of simple fragments: 1, 10 and 100 for `benchmark1`, `benchmark10` and `benchmark100` respectively. Enabling validation cache improves execution time:
- `benchmark1` from 4.222ms to 2.886ms 
- `benchmark10` from 14.984ms  to 4.429ms 
- `benchmark100` from 189.641ms to 19.783ms