@k0ka 2024-09-19
Hello @spawnia,

do you have time to check this PR? 
It significantly improves speed and is disabled by default. So it won't break backward compatibility and we can gradually test it on real loads.