@spawnia 2024-09-20
Released with https://github.com/nuwave/lighthouse/releases/tag/v6.45.0.